Moved topic to datatypes package
