+++ src/main/java/model/FXModel.java
@@ -5,8 +5,8 @@ import api.TopicListener;
 import data.MessagePublication;
 import data.MessagePublish;
 import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.collections.ObservableMap;
+import model.datatypes.Topic;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import utils.ArrayUtils;
